feat(tinyvue):  views/list 拆分编辑信息的组件
